% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{beam_beam}
\alias{beam_beam}
\title{beam_beam}
\usage{
beam_beam(kicked, kickers, sim)
}
\value{
data.frame summary, including corrections to overlap integrals
}
\description{
Simulates bunch particles in an accelerator influenced by the
             electromagnetic interaction with another bunch ("beam-beam"
             effect). Reports corrections to the bunches overlap integral.
}
\author{
Vladislav BALAGURA <balagura@cern.ch>
}
