% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{package}
\name{BxB}
\alias{BxB}
\title{B*B, beam-beam simulation for vdM scans at LHC}
\description{
"beam_beam" function simulates bunch particles in an
             accelerator influenced by the electromagnetic force of
             another bunch ("beam-beam" effect). Reports corrections
             to the bunches overlap integral. See "?beam_beam" for details.
}
\author{
Vladislav BALAGURA <balagura@cern.ch>
}
